TINY-9670: Move iframe init code sideways from InitContentBody to InitIframe (#8531)
Related Ticket: TINY-9670

Description of Changes:
* While explaining the Init process to @kemister85, I realised that the iframe-specific things `InitContentBody` does no longer need to be there. The `skipWrite` parameter is never used.
* So I moved all the code to `InitIframe`, which simplifies the explanation considerably.

Pre-checks:
* [x] ~Changelog entry added~
* [x] ~Tests have been added (if applicable)~
* [x] Branch prefixed with `feature/`, `hotfix/` or `spike/`

Review:
* [x] Milestone set
* [x] ~Docs ticket created (if applicable)~

GitHub issues (if applicable):
